Move AI trace log to default plugin
# Conflicts:
#	plugin-default-impl/nacos-default-plugin-all/pom.xml
#	plugin-default-impl/pom.xml
