+++ plugin-default-impl/pom.xml
@@ -35,6 +35,7 @@
         <module>nacos-ldap-auth-plugin</module>
         <module>nacos-default-ai-pipeline-plugin</module>
         <module>nacos-default-ai-importer-plugin</module>
+        <module>nacos-default-ai-trace-plugin</module>
         <module>nacos-default-datasource-plugin</module>
         <module>nacos-default-plugin-all</module>
         <module>nacos-oidc-auth-plugin</module>
